[MIG]  stock_picking_return_restricted_qty: migrate to v19
Updates the module to Odoo 19.0, refactoring the return picking logic
to use modern Odoo patterns, separating wizard classes into distinct files,
and ensuring compatibility with the latest framework standards. It maintains
the restriction on return quantities and adapts tests and code structure
for maintainability.
