+++ stock_picking_return_restricted_qty/README.rst
@@ -2,7 +2,7 @@
 Stock Picking Return Restricted Qty
 ===================================
 
-.. 
+..
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
