+++ stock_picking_return_restricted_qty/__manifest__.py
@@ -4,7 +4,7 @@
 {
     "name": "Stock Picking Return Restricted Qty",
     "summary": "Restrict the return to delivered quantity",
-    "version": "18.0.1.0.0",
+    "version": "19.0.1.0.0",
     "license": "AGPL-3",
     "author": "Tecnativa,Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/stock-logistics-workflow",
