+++ stock_picking_return_restricted_qty/wizard/__init__.py
@@ -1,3 +1,3 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import stock_picking_return
+from . import return_picking, return_picking_line
